Conflicts:
	Makefile
	modules/pico_dhcp_client.c
	modules/pico_dns_client.c
	modules/pico_http_server.c
	modules/pico_http_util.c
	modules/pico_ipfilter.c
	modules/pico_mm.c
	stack/pico_device.c

+++ modules/pico_icmp4.c
@@ -236,7 +236,7 @@ static inline void send_ping(struct pico_icmp4_ping_cookie *cookie)
     pico_icmp4_send_echo(cookie);
     cookie->timestamp = pico_tick;
     pico_timer_add((uint32_t)cookie->timeout, ping_timeout, cookie);
-    if (cookie->seq < cookie->count)
+    if (cookie->seq < (uint16_t)cookie->count)
         pico_timer_add((uint32_t)cookie->interval, next_ping, cookie);
 }
 
@@ -246,7 +246,7 @@ static void next_ping(pico_time now, void *arg)
     IGNORE_PARAMETER(now);
 
     if(pico_tree_findKey(&Pings, cookie)) {
-        if (cookie->seq < cookie->count) {
+        if (cookie->seq < (uint16_t)cookie->count) {
             newcookie = PICO_ZALLOC(sizeof(struct pico_icmp4_ping_cookie));
             if (!newcookie)
                 return;
